VIOLENCE AGAINST WOMEN ARE CALL SYSTEMS
The name of the program is stop violence against women call systems. WS for short. The aim of this program is to detect violence against women as soon as possible and to send a police team to the address where the violence took place. The information of women who have searched before is stored in the database